Update to v0.2.7 Consult CHANGELOG.md
- Fix sitemap.xml listing items and exposing content that is only available to logged in accounts, and adjust listing so pages/content listed as noindex do not get listed.
- Add more WordPress filtering methods for known attack attempts.
- Add new option to Visitor Tracker to clear entries via Search Query.
- Update Video Embedding within the content parser, as well as Summernote when adding video content.
- Add a check for when an outside source is looking for ads.txt or sellers.json to avoid having those URL's added to the tracking database. (We may visit this as an option, as this is geared mostly towards when advertising is placed on pages, such as Google's AdSense.)
- Add enterkeyhint attributes to form input element to improve mobile usage.
- Remove Title and trailing / from RSS feed link meta tag.
- Fix RSS Feed output listing items and exposing content that is only available to logged in accounts, and adjust listing so pages/content listed as noindex do not get listed.
- Add regex to remove extraneous white space from RSS feed description text.
- Fix Facebook and Twitter Share Icons.
- Fix missing Twiiter Card information.
- Fix .htaccess URL 301 Redirect to properly resolve to https and without www
- Add response to access to security.txt saying to report security issues to the github repository.
- Add new headers for Client Hints directives.
- Improve how tracking site visitors works, now records more browsers, operating systems, screen width and Telephone Call Clicks.
- Fix Item Count error in SQL when gathering content.
- Add Calls count for current month to Dashboard.
- Update Google Analytics Embed Code to self hosted GA4, self hosting helps speed up page speed.
- Improve Schema output, now includes dynamic Organization rather than static.
- Add Google Tag Manager and ID Entry.
- Adjust Rank for Wholesale (Bronze and Silver were reversed), and add Standard Wholesale.
- Add Free SEO Course links on Dashboard.
- Fix issue when updating postcode values, which would not update, and throw error if null/empty value used.

+++ core/purge.php
@@ -19,7 +19,7 @@ $tbl=isset($_POST['t'])?filter_input(INPUT_POST,'t',FILTER_SANITIZE_STRING):filt
 $col=isset($_POST['c'])?filter_input(INPUT_POST,'c',FILTER_SANITIZE_STRING):filter_input(INPUT_GET,'c',FILTER_SANITIZE_STRING);
 $uid=isset($_SESSION['uid'])?$_SESSION['uid']:0;
 $el='l_';
-if($id!=0&&$tbl!='logs'&&$tbl!='livechat'&&$tbl!='forumCategory'&&$tbl!='forumTopics'&&$tbl!='pageviews'&&$tbl!='playlist'){
+if($id!=0&&$tbl!='logs'&&$tbl!='livechat'&&$tbl!='forumCategory'&&$tbl!='forumTopics'&&$tbl!='pageviews'&&$tbl!='playlist'&&$tbl!='findtracker'){
   $s=$db->prepare("SELECT * FROM `".$prefix.$tbl."` WHERE `id`=:id");
   $s->execute([':id'=>$id]);
   $r=$s->fetch(PDO::FETCH_ASSOC);
@@ -47,6 +47,11 @@ if($id==0&&$tbl=='logs'){
   $q=$db->query("DELETE FROM `".$prefix."logs`");
   $q->execute();
   $id='timeline';
+}
+if($tbl=='findtracker'){
+  $q=$db->prepare("DELETE FROM `".$prefix."tracker` WHERE `urlDest`=:id");
+  $q->execute([':id'=>'%'.strtolower($id).'%']);
+  $id='tracker';
 }
 if($id==0&&$tbl=='tracker'){
   $q=$db->query("DELETE FROM `".$prefix."tracker`");
@@ -81,7 +86,7 @@ if($tbl=='playlist'){
   $s=$db->prepare("DELETE FROM `".$prefix."choices` WHERE `id`=:id");
   $s->execute([':id'=>$id]);
 }
-if($id!=0&&$id!='activity'&&$tbl!='pageviews'&&$tbl!='playlist'){
+if($id!=0&&$id!='activity'&&$tbl!='pageviews'&&$tbl!='playlist'&&$tbl!='findtracker'){
   $q=$db->prepare("DELETE FROM `".$prefix.$tbl."` WHERE `id`=:id");
   $q->execute([':id'=>$id]);
   if($tbl=='media')$el='media_items_';
